refactor(shelter): 보호소 정보 Fetch 훅 수정
